+++ test/test_detailed_hydro.jl
@@ -119,7 +119,6 @@ end
         @test outputs(res, water) == 1
         @test node_data(res) == res_data
         @test capacity(gate) == FixedProfile(100)
-        @test all(capacity(gate, t, water) == capacity(gate, t) for t ∈ 𝒯)
         @test all(capacity(gate, t) == 100 for t ∈ 𝒯)
         @test opex_var(gate) == FixedProfile(0)
         @test all(opex_var(gate, t) == 0 for t ∈ 𝒯)
@@ -214,6 +213,7 @@ end
                 value.(m[:sink_deficit][sink, t]) * deficit_penalty(sink, t) * duration(t) +
                 value.(m[:rsv_penalty_up][res, t, water]) * 10 * duration(t)
         for t ∈ 𝒯)
+        @test objective_value(m) ≈-100
     end
 
     @testset "Hydro reservoir - Hard MinSchedule and soft MaxSchedule" begin
@@ -266,6 +266,7 @@ end
                 sum(value.(m[:rsv_penalty_down][res, t, water]) * penalty_cost * duration(t) for t ∈ t_inv)
         for t_inv ∈ 𝒯ⁱⁿᵛ)
         @test objective_value(m) ≈ -sum(value.(m[:opex_var][res, t_inv]) for t_inv ∈ 𝒯ⁱⁿᵛ)
+        @test objective_value(m) ≈ -1140
     end
 
     @testset "Hydro reservoir - Hard EqualSchedule" begin
@@ -294,14 +295,13 @@ end
         @test isempty(m[:rsv_penalty_up])
         @test isempty(m[:rsv_penalty_down])
 
-        # Test that there are no violations and the storage level variables are fixed
-        # - build_hydro_reservoir_vol_constraints(m, n::HydroReservoir, c::ScheduleConstraint{EqualSchedule}, 𝒯)
+        # Test that there are no violations
+        # - EMB.constraints_ext_data(m, n::HydroNode, 𝒯, 𝒫, modeltype::EnergyModel, data::ScheduleConstraint{EqualSchedule})
         @test all(
             value.(m[:stor_level][res, t]) ≈ sched_profile[t] * capacity(level(res), t)
         for t ∈ 𝒯)
         prof = OperationalProfile([15, 0, 0, 5])
         @test all(value.(m[:sink_deficit][sink, t]) ≈ prof[t] for t ∈ 𝒯)
-        @test all(is_fixed(m[:stor_level][res, t]) for t ∈ 𝒯)
     end
 
     @testset "Hydro reservoir - Soft EqualSchedule" begin
@@ -326,13 +326,20 @@ end
         𝒯ⁱⁿᵛ = strategic_periods(𝒯)
         res, gate, sink = get_nodes(case)[[1, 2, 3]]
 
+        # Test that the variable extraction functions are working
+        @test EMRP.get_var_pen_up(m, res, 𝒯, res_data[1]) ==
+            m[:rsv_penalty_up][res, :, water]
+        @test EMRP.get_var_pen_down(m, res, 𝒯, res_data[1]) ==
+            m[:rsv_penalty_down][res, :, water]
+        @test EMRP.get_var_schedule(m, res, 𝒯, res_data[1]) == m[:stor_level][res, :]
+
         # Test that the penalty variables are created
         # - EMB.variables_node(m, 𝒩::Vector{<:HydroReservoir}, 𝒯, modeltype::EnergyModel)
         @test !isempty(m[:rsv_penalty_up])
         @test !isempty(m[:rsv_penalty_down])
 
         # Test that the violations are correctly calculated
-        # - build_hydro_reservoir_vol_constraints(m, n::HydroReservoir, c::ScheduleConstraint{EqualSchedule}, 𝒯)
+        # - EMB.constraints_ext_data(m, n::HydroNode, 𝒯, 𝒫, modeltype::EnergyModel, data::ScheduleConstraint{EqualSchedule})
         @test all(
             value.(m[:stor_level][res, t] - m[:rsv_penalty_down][res, t, water]) ≤
                 sched_profile[t] * capacity(level(res), t)
@@ -354,6 +361,7 @@ end
                 penalty_cost * duration(t) for t ∈ t_inv)
         for t_inv ∈ 𝒯ⁱⁿᵛ)
         @test objective_value(m) ≈ -sum(value.(m[:opex_var][res, t_inv]) for t_inv ∈ 𝒯ⁱⁿᵛ)
+        @test objective_value(m) ≈ -400
     end
 
     @testset "Gate - Hard EqualSchedule" begin
@@ -386,10 +394,9 @@ end
         @test isempty(m[:gate_penalty_up])
         @test isempty(m[:gate_penalty_down])
 
-        # Test that there are no violations and the otflow variables are fixed
-        # - build_schedule_constraint(m, n::Union{HydroGate, HydroUnit}, c::ScheduleConstraint{EqualSchedule}, 𝒯::TimeStructure, p::ResourceCarrier)
+        # Test that there are no violations
+        # - EMB.constraints_ext_data(m, n::HydroNode, 𝒯, 𝒫, modeltype::EnergyModel, data::ScheduleConstraint{EqualSchedule})
         @test all(gate_flow[t] ≈ schedule_profile[t] * capacity(gate, t) for t ∈ 𝒯)
-        @test all(is_fixed(m[:flow_out][gate, t, water]) for t ∈ 𝒯)
     end
 
     @testset "Gate - Soft EqualSchedule, varying flags" begin
@@ -416,21 +423,28 @@ end
         res, gate, sink = get_nodes(case)[[1, 2, 3]]
         gate_flow = value.(m[:flow_out][gate, :, water])
 
-        # Test that the outflow is high when there are no penalties
-        prof = OperationalProfile([5, 10, 10, 22.5])
-        @test all(gate_flow[t] ≈ prof[t] for t ∈ 𝒯)
+        # Test that the variable extraction functions are working
+        @test EMRP.get_var_pen_up(m, gate, 𝒯, gate_data[1]) ==
+            m[:gate_penalty_up][gate, :, water]
+        @test EMRP.get_var_pen_down(m, gate, 𝒯, gate_data[1]) ==
+            m[:gate_penalty_down][gate, :, water]
+        @test EMRP.get_var_schedule(m, gate, 𝒯, gate_data[1]) == m[:flow_out][gate, :, water]
 
         # Test that the penalty variables are created and non-empty
         # - EMB.variables_node(m, 𝒩::Vector{HydroGate}, 𝒯, modeltype::EnergyModel)
         @test !isempty(m[:gate_penalty_up])
         @test !isempty(m[:gate_penalty_down])
 
+        # Test that the outflow is high when there are no penalties
+        prof = OperationalProfile([5, 10, 10, 22.5])
+        @test all(gate_flow[t] ≈ prof[t] for t ∈ 𝒯)
+
         # Test that the constraint is enforced
         @test all(iszero(value.(m[:gate_penalty_up][gate, t, water])) for t ∈ 𝒯 if flags[t])
         @test all(iszero(value.(m[:gate_penalty_down][gate, t, water])) for t ∈ 𝒯 if flags[t])
 
         # Test that the schedule values are used, when the flag is set
-        # - build_schedule_constraint(m, n::Union{HydroGate, HydroUnit}, c::ScheduleConstraint{EqualSchedule}, 𝒯::TimeStructure, p::ResourceCarrier)
+        # - EMB.constraints_ext_data(m, n::HydroNode, 𝒯, 𝒫, modeltype::EnergyModel, data::ScheduleConstraint{EqualSchedule})
         @test all(gate_flow[t] ≈ schedule_profile[t]*capacity(gate, t) for t ∈ 𝒯 if flags[t])
     end
 
@@ -481,6 +495,7 @@ end
                     scale_op_sp(t_inv, t) * penalty_cost[t] * value.(m[:gate_penalty_down][gate, t, water])
                 for t ∈ t_inv)
         for t_inv ∈ 𝒯ⁱⁿᵛ)
+        @test all(value.(m[:opex_var][gate, t_inv]) ≈ 60 for t_inv ∈ 𝒯ⁱⁿᵛ)
     end
 
     @testset "Gate - Soft MinSchedule and hard MaxSchedule, varying penalty" begin
@@ -530,6 +545,7 @@ end
                     scale_op_sp(t_inv, t) * penalty_cost[t] * value.(m[:gate_penalty_up][gate, t, water])
                 for t ∈ t_inv)
         for t_inv ∈ 𝒯ⁱⁿᵛ)
+        @test all(value.(m[:opex_var][gate, t_inv]) ≈ 36 for t_inv ∈ 𝒯ⁱⁿᵛ)
     end
 end
 
@@ -616,8 +632,6 @@ end
         # Test the EMB utility functions
         @test capacity(gen) == FixedProfile(20)
         @test all(capacity(gen, t) == 20 for t ∈ 𝒯)
-        @test all(capacity(gen, t, power) == capacity(gen, t) for t ∈ 𝒯)
-        @test all(capacity(gen, t, water) == capacity(gen, t) * 1.1 for t ∈ 𝒯)
         @test opex_var(gen) == FixedProfile(0)
         @test all(opex_var(gen, t) == 0 for t ∈ 𝒯)
         @test opex_fixed(gen) == FixedProfile(0)
@@ -735,15 +749,54 @@ end
         @test !isempty(m[:discharge_segment][gen, :, :])
         @test all(length(m[:discharge_segment][gen, t, :]) == 2  for t ∈ 𝒯)
 
-        # Test that the penalty variables are not created
-        # - EMB.variables_node(m, 𝒩::Vector{HydroUnit}, 𝒯, modeltype::EnergyModel)
-        @test isempty(m[:gen_penalty_up])
-        @test isempty(m[:gen_penalty_down])
-
-        # Test that there are no violations and the outflow variables are fixed when required
-        # - build_schedule_constraint(m, n::Union{HydroGate, HydroUnit}, c::ScheduleConstraint{EqualSchedule}, 𝒯::TimeStructure, p::ResourceCarrier)
+        # Test that there are no violations
+        # - EMB.constraints_ext_data(m, n::HydroNode, 𝒯, 𝒫, modeltype::EnergyModel, data::ScheduleConstraint{EqualSchedule})
         @test all(gen_out[t, power] ≈ schedule_profile[t] * capacity(gen, t) for t ∈ 𝒯 if schedule_flag[t])
-        @test all(is_fixed(m[:flow_out][gen, t, power]) for t ∈ 𝒯 if schedule_flag[t])
+    end
+
+    @testset "Soft EqualSchedule for power" begin
+        # Modify the input data
+        schedule_profile = OperationalProfile(0.8 * ones(4))
+        schedule_flag = OperationalProfile([false, false, true, true])
+        data = [
+            ScheduleConstraint{EqualSchedule}(
+                power,
+                schedule_profile,   # value
+                schedule_flag,      # flag
+                FixedProfile(200),  # penalty
+            )
+        ]
+        profit = OperationalProfile(-[50, 50, 10, 10])
+
+        # Create and solve the model
+        case, modeltype = gen_test_case(; data, profit)
+        m = EMB.run_model(case, modeltype, OPTIMIZER)
+
+        # Extract the data
+        𝒯 = get_time_struct(case)
+        𝒯ⁱⁿᵛ = strategic_periods(𝒯)
+        gate, sink, gen = get_nodes(case)[2:4]
+        pq_val = EMRP.pq_curve(gen)
+        gen_out = value.(m[:flow_out][gen, :, :])
+
+        # Test that the penalty variables are created including discharge_segments
+        # - EMB.variables_node(m, 𝒩::Vector{HydroUnit}, 𝒯, modeltype::EnergyModel)
+        @test !isempty(m[:gen_penalty_up])
+        @test !isempty(m[:gen_penalty_down])
+        @test !isempty(m[:discharge_segment][gen, :, :])
+        @test all(length(m[:discharge_segment][gen, t, :]) == 2  for t ∈ 𝒯)
+
+        # Test that the variable extraction functions are working
+        @test EMRP.get_var_pen_up(m, gen, 𝒯, data[1]) ==
+            m[:gen_penalty_up][gen, :, power]
+        @test EMRP.get_var_pen_down(m, gen, 𝒯, data[1]) ==
+            m[:gen_penalty_down][gen, :, power]
+        @test EMRP.get_var_schedule(m, gen, 𝒯, data[1]) == m[:flow_out][gen, :, power]
+
+        # Test that there are no violations and the outflow variables are not fixed
+        # - EMB.constraints_ext_data(m, n::HydroNode, 𝒯, 𝒫, modeltype::EnergyModel, data::ScheduleConstraint{EqualSchedule})
+        @test all(gen_out[t, power] ≈ schedule_profile[t] * capacity(gen, t) for t ∈ 𝒯 if schedule_flag[t])
+        @test all(!is_fixed(m[:flow_out][gen, t, power]) for t ∈ 𝒯)
     end
 
     @testset "Soft MinSchedule for water" begin
@@ -776,8 +829,8 @@ end
         @test !isempty(m[:gen_penalty_down])
 
         # Test that outflow is constrained due to the large penalty
-        # - build_schedule_constraint(m, n::Union{HydroGate, HydroUnit}, c::ScheduleConstraint{EqualSchedule}, 𝒯::TimeStructure, p::ResourceCarrier)
-        @test all(gen_out[t, water] ≈ schedule_profile[t] * capacity(gen, t, water) for t ∈ 𝒯)
+        # - EMB.constraints_ext_data(m, n::HydroNode, 𝒯, 𝒫, modeltype::EnergyModel, data::ScheduleConstraint{EqualSchedule})
+        @test all(gen_out[t, water] ≈ schedule_profile[t] * capacity(gen, t) * 1.1 for t ∈ 𝒯)
     end
 end
 
@@ -892,7 +945,6 @@ end
         return case, modeltype
     end
 
-
     @testset "Utlities" begin
         # Create the model and extract the data
         val = OperationalProfile(0.1 * ones(4))
@@ -924,8 +976,6 @@ end
         # Test the EMB utility functions
         @test capacity(pump) == FixedProfile(30)
         @test all(capacity(pump, t) == 30 for t ∈ 𝒯)
-        @test all(capacity(pump, t, power) == capacity(pump, t) for t ∈ 𝒯)
-        @test all(capacity(pump, t, water) == capacity(pump, t) * 2/3 for t ∈ 𝒯)
         @test opex_var(pump) == FixedProfile(0)
         @test all(opex_var(pump, t) == 0 for t ∈ 𝒯)
         @test opex_fixed(pump) == FixedProfile(0)
@@ -1049,8 +1099,57 @@ end
         @test all(length(m[:discharge_segment][pump, t, :]) == 2  for t ∈ 𝒯)
 
         # Test that there are no violations on the scheduling constraints
-        # - build_schedule_constraint(m, n::Union{HydroGate, HydroUnit}, c::ScheduleConstraint{EqualSchedule}, 𝒯::TimeStructure, p::ResourceCarrier)
-        @test all(gen_out[t, water] ≥ 0.6 * capacity(gen, t, water) for t ∈ 𝒯 if gen_flag[t])
-        @test all(pump_out[t, water] ≥ 0.4 * capacity(pump, t, water) for t ∈ 𝒯 if pump_flag[t])
+        # - EMB.constraints_ext_data(m, n::HydroNode, 𝒯, 𝒫, modeltype::EnergyModel, data::ScheduleConstraint{EqualSchedule})
+        @test all(gen_out[t, water] ≥ 0.6 * capacity(gen, t) for t ∈ 𝒯 if gen_flag[t])
+        @test all(pump_out[t, water] ≥ 0.4 * capacity(pump, t)*2/3 for t ∈ 𝒯 if pump_flag[t])
+    end
+    @testset "Soft MinSchedule for water" begin
+        # Modify the input data
+        gen_flag = OperationalProfile([true, false, false, false])
+        gen_data = [ScheduleConstraint{MinSchedule}(
+            water,
+            FixedProfile(0.6),  # value
+            gen_flag,           # flag
+            FixedProfile(Inf),  # penalty
+        )]
+        pump_flag = OperationalProfile([false, true, false, false])
+        pump_data = [ScheduleConstraint{EqualSchedule}(
+            water,
+            FixedProfile(0.4),  # value
+            pump_flag,          # flag
+            FixedProfile(50),  # penalty
+        )]
+
+        # Create and solve the model
+        case, modeltype = pump_test_case(; gen_data, pump_data)
+        m = EMB.run_model(case, modeltype, OPTIMIZER)
+
+        # Extract the pump_data
+        𝒯 = get_time_struct(case)
+        𝒯ⁱⁿᵛ = strategic_periods(𝒯)
+        gen, pump = get_nodes(case)[3:4]
+        pq_val = EMRP.pq_curve(pump)
+        gen_out = value.(m[:flow_out][gen, :, :])
+        pump_in = value.(m[:flow_in][pump, :, :])
+        pump_out = value.(m[:flow_out][pump, :, :])
+
+        # Test that the penalty variables are created, including discharge_segments
+        # - EMB.variables_node(m, 𝒩::Vector{HydroUnit}, 𝒯, modeltype::EnergyModel)
+        @test !isempty(m[:gen_penalty_up])
+        @test !isempty(m[:gen_penalty_down])
+        @test !isempty(m[:discharge_segment][pump, :, :])
+        @test all(length(m[:discharge_segment][pump, t, :]) == 2  for t ∈ 𝒯)
+
+        # Test that the variable extraction functions are working
+        @test EMRP.get_var_pen_up(m, pump, 𝒯, pump_data[1]) ==
+            m[:gen_penalty_up][pump, :, water]
+        @test EMRP.get_var_pen_down(m, pump, 𝒯, pump_data[1]) ==
+            m[:gen_penalty_down][pump, :, water]
+        @test EMRP.get_var_schedule(m, pump, 𝒯, pump_data[1]) == m[:flow_in][pump, :, water]
+
+        # Test that there are no violations on the scheduling constraints
+        # - EMB.constraints_ext_data(m, n::HydroNode, 𝒯, 𝒫, modeltype::EnergyModel, data::ScheduleConstraint{EqualSchedule})
+        @test all(gen_out[t, water] ≥ 0.6 * capacity(gen, t) * 2/3 for t ∈ 𝒯 if gen_flag[t])
+        @test all(pump_out[t, water] ≥ 0.4 * capacity(pump, t) * 2/3 for t ∈ 𝒯 if pump_flag[t])
     end
 end